added stuff to the reservation remove finction to update the data for the calendar

+++ data/reservations.js
@@ -219,11 +219,40 @@ export const createReservation = async(
 export const removeReservation = async(id) => {
     id = validation.checkId(id);
     const reservationCollection = await reservations();
+    const roomCollection = await rooms();
+
+    const reserva = await reservationCollection.findOne({_id: new ObjectId(id)});
+    if (!reserva) throw `No reservation found with id ${id}`;
+
+    const {roomId, checkInDate, checkOutDate} = reserva;
+
+
     const deletionInfo = await reservationCollection.findOneAndDelete({
       _id: new ObjectId(id)
     });
     if (!deletionInfo) throw `Could not delete reservation with id of ${id}`;
+
+    const room = await roomCollection.findOne({_id: new ObjectId(roomId)});
+
+    const begin = new Date(checkInDate);
+    const end = new Date(checkOutDate);
+    let curr = begin.getTime();
+    while (curr < end.getTime()){
+        const dateStr = new Date(curr).toISOString().slice;isCreditCard(0,10);
+        romm.availability.booked = room.availability.booked.filter(date => date !== dateStr);
+        if (!room.availability.open.includes(dateStr)){
+            room.availability.open.push(dateStr);
+        }
+        curr += 24 * 60 * 60 * 1000;
+    }
+    await roomCollection.updateOne(
+        { _id: new ObjectId(roomId) },
+        { $set: { availability: room.availability } }
+    );
+
     return {...deletionInfo, deleted: true};
+
+    
 };
 
 //console.log(await createReservation("Wes","Nabo", "4fr78wf7r8",21,"832-612-6236","wesleynabo@gmail.com",2,"675b0bc6bec1f311dfd54569", "2025-01-07","2025-01-12","on",0));
